+++ src/App.js
@@ -11,7 +11,8 @@ import instagram from './img/rede__instagram.svg'
 import youtube from './img/rede__youtube.svg'
 import twitter from './img/rede__twitter.svg'
 import whatsapp from './img/rede__whatsapp.svg'
-
+import TelaCadastro from "./components/TelaCadastro"
+import TelaSelecao from "./components/TelaSelecao"
 import LogoVertical from "./img/LogoVertical.png"
 
 
@@ -73,91 +74,6 @@ const FiltrosSacola = styled.div`
 	grid-column-end: 4;
 `
 
-const BoasVindas = styled.h2`
-	display: flex;
-	justify-content: center;
-	align-items: flex-end;
-`
-
-const TextoIntroducao = styled.h3`
-	text-align: center;
-`
-
-const BotoesInicio = styled.button`
-	width: 95px;
-	height: 45px;
-	margin: 6px 15px;
-	border-radius: 8%;
-	background-color: #59D9D9;
-`
-
-class TelaSelecao extends React.Component {
-	
-	render() {
-
-		return(
-		<div>
-		 <BoasVindas>Bem vindo(a) ao Elo4!</BoasVindas>
-          <TextoIntroducao>A maior rede de compra e venda de produtos artesanais do Brasil!</TextoIntroducao>
-          <TextoIntroducao>Escolha o que deseja fazer:</TextoIntroducao>
-          <div>
-			<BotoesInicio>Quero Comprar!</BotoesInicio>
-            <BotoesInicio>Quero Vender!</BotoesInicio>
-          </div>
-		</div>	
-		)
-	}
-
-}
-
-class TelaCadastro extends React.Component {
-	render() {
-
-		return(
-		<div>
-			<h2>Olá lojista!</h2>
-			<h2>Aqui é o seu espaço!</h2>
-			<h3>Cadastre aqui os dados do seu produto.</h3>
-			<input placeholder="Digite o nome do produto"></input>
-			<input placeholder="Escreve a descrição do produto"></input>
-			<select>
-				<option value="">Escola a categoria:</option>
-				<option value="festas">Artigos para festa</option>
-				<option value="decoracao">Objetos de Decoração</option>
-				<option value="bijuteria">Bijuteria</option>
-				<option value="religiosos">Produtos Religiosos</option>
-				<option value="enxovais">Enxovais</option>
-				<option value="ecofriendly">Acessórios Eco-friendly</option>
-			</select>
-			<select>
-				<option value="">Escola o métdo de pagamento:</option>
-				<option value="debito">Cartão de Débito</option>
-				<option value="credito">Cartão de Crédito</option>
-				<option value="boleto">Boleto Bancário</option>
-			</select>
-			<input placeholder="Digite o preço em R$"></input>
-			<select>
-				<option value="">Escolha o número máximo de parcelas:</option>
-				<option value="1">Á vista</option>
-				<option value="2">2 vezes</option>
-				<option value="3">3 vezes</option>
-				<option value="4">4 vezes</option>
-				<option value="5">5 vezes</option>
-				<option value="6">6 vezes</option>
-				<option value="7">7 vezes</option>
-				<option value="8">8 vezes</option>
-				<option value="9">9 vezes</option>
-				<option value="10">10 vezes</option>
-				<option value="11">11 vezes</option>
-				<option value="12">12 vezes</option>
-			</select>
-			<input placeholder="Digite a url da imagem do produto"></input>
-			<button>Cadastrar</button><button>Voltar</button>
-		</div>	
-		)
-	}
-
-}
 
 
 //let telaExibida = {<TelaSelecao/>}
@@ -176,6 +92,7 @@ class App extends React.Component {
                 </ContainerCentral>
             <FiltrosSacola>
                 <Filtros/>
+				<Sacola/>
             </FiltrosSacola>
             <Footer>
               <Icones>
